Add a link from the framesplit to the build screen
Hidden, but magically appearing on hover (over a transparent area)
so it doesn't interfer visually at all, but is relatively easy
to get to.

Better than having to edit the URL.

If it's loaded without URLs, then the edit link is shown
unconditionally, which is actually an improvement over the
previous (completely) blank default.